pyproject.toml
[project]
name = "llm-orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Azzarnuji",email = "azzarnuzy93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-genai (>=1.29.0,<2.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
